project_#4_Environmental Domain LLM Fine-tuning/requirements.txt
# =============================================================================
# Environmental Domain LLM Fine-tuning - Dependencies
# =============================================================================

# Core ML Framework
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Transformers & PEFT
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
safetensors>=0.4.0

# Datasets
datasets>=2.16.0
huggingface_hub>=0.20.0

# Training Utilities
trl>=0.7.0
wandb>=0.16.0
tensorboard>=2.15.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0

# Tokenization
tiktoken>=0.5.0
sentencepiece>=0.1.99

# Evaluation
evaluate>=0.4.0
rouge_score>=0.1.2
nltk>=3.8.0
sacrebleu>=2.4.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
omegaconf>=2.3.0

# Inference & Serving
fastapi>=0.109.0
uvicorn>=0.25.0
gradio>=4.10.0

# Export
llama-cpp-python>=0.2.0
onnx>=1.15.0
optimum>=1.15.0

# Utilities
tqdm>=4.66.0
rich>=13.7.0
loguru>=0.7.2
click>=8.1.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.1.0

# Testing
pytest>=7.4.0

# Development
black>=23.12.0
isort>=5.13.0
